Organize react component files
